pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "face-recognition>=1.3.0",
    "face-recognition-models",
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "setuptools<71.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pymongo>=4.7.0",
    "mysql-connector-python>=9.0.0",
    "pandas>=2.2.0",
    "requests>=2.32.0",
    "python-multipart>=0.0.9",
]

[tool.uv.sources]
face-recognition-models = { git = "https://github.com/ageitgey/face_recognition_models" }
